Move things around, make ble fully in network
