+++ src/network/ble/gadgetbridge/gadgetbridge.cpp
