+++ src/network/ble/gadgetbridge/gadgetbridge.h
